JUnit setup and TestValidWord partial implementation
JUnit 5 has been setup and all necessary file structures to run tests
are in place. KeyboardButtonActionTest file has been created to house
tests on the KeyboardButtonAction file, specifically testing for if the
file correctly doesn't allow a user to enter an invalid word. The test
has been partially implemented - a reader reads through the usa.txt file
and can correctly determine if an input word is in that file. However,
work still needs to be done to simulate a real game with a real input.