TestValidWord
This version fully implements a test that tests if invalid words are
able to be guessed. It does this by creating a dummy WordleModel, taking
it's current row (-1), simulating a user guessing "abcde" and hitting
the enter button, then checking the current row after the guess. The
correct functionality of the program would be to block the invalid guess
- preventing the row count from incrementing. Therefore, the test only
passes when the starting row and ending rows are the same. Currently,
the test fails since the bug fix has not yet been implemented.