checkValidWord implementation (with WordleModel)
This update fixes the bug of a user being able to guess an invalid word
through a new method called checkValidWord. checkValidWord reads through
the usa.txt file of valid words and checks if the input guess matches
any of them. If so, the method returns true and the "Enter" case
continues as normal in the actionPerformed method. If not, the method
returns false and the "Enter" case breaks and no change is made to the
model. Previously, this update was pushed to the git repository, but the
WordleModel file was assumed unchanged - this commit adds the changed
WordleModel file to the repo.