WordleModelTest
This update creates a new JUnit test file that contains all methods for
testing the difficulty of the game. The game's difficulty settings have
not yet been implemented in the model, so most tests currently fail.
Also, most tests rely on testing a setDifficulty() function that does
not yet exist for WordleModel, so many of these tests currently produce
errors.