Difficulty Implementation
This large update adds a difficulty menu to the menu bar. When clicked,
the menu drops down and the user can select "Easy", "Medium", or "Hard".
After making their selection, the current instance of the game being
played is aborted and a new game with the selected difficulty begins.
The JUnit test file for these changes has also been tweaked to stay
within the scope of WordleModel - reseting the game is no longer tested
since it is handled by WordleFrame after the difficulty has been
changed.